requirements.txt
Flask
PyPDF2
pytesseract
pillow
textblob
